+++ src/Irooon.Core/CodeGen/CodeGenerator.cs
@@ -69,6 +69,7 @@ public class CodeGenerator
             MemberAssignExpr e => GenerateMemberAssignExpr(e),
             StringInterpolationExpr e => GenerateStringInterpolationExpr(e),
             TryExpr e => GenerateTryExpr(e),
+            RangeExpr e => GenerateRangeExpr(e),
             _ => throw new NotImplementedException($"Unknown expression type: {expr.GetType()}")
         };
     }
@@ -88,7 +89,7 @@ public class CodeGenerator
             VarStmt s => GenerateVarStmt(s),
             ExprStmt s => GenerateExprStmt(s),
             ReturnStmt s => GenerateReturnStmt(s),
-            WhileStmt s => GenerateWhileStmt(s),
+            ForStmt s => GenerateForStmt(s),
             ForeachStmt s => GenerateForeachStmt(s),
             BreakStmt s => GenerateBreakStmt(s),
             ContinueStmt s => GenerateContinueStmt(s),
@@ -454,55 +455,6 @@ public class CodeGenerator
     }
 
     /// <summary>
-    /// while文の変換
-    /// 仕様（expression-tree-mapping.md セクション8）:
-    /// Expression.Loop + LabelBreak
-    /// while全体の値は null
-    /// </summary>
-    private ExprTree GenerateWhileStmt(WhileStmt stmt)
-    {
-        var breakLabel = ExprTree.Label($"whileBreak_{_labelCounter}");
-        var continueLabel = ExprTree.Label($"whileContinue_{_labelCounter++}");
-
-        // ループラベルをスタックにプッシュ
-        _loopLabels.Push((breakLabel, continueLabel));
-
-        var condExpr = GenerateExpression(stmt.Condition);
-        var bodyExpr = GenerateStatement(stmt.Body);
-
-        // ループラベルをポップ
-        _loopLabels.Pop();
-
-        // IsTruthy で真偽値判定
-        var truthyCall = ExprTree.Call(
-            typeof(RuntimeHelpers),
-            nameof(RuntimeHelpers.IsTruthy),
-            null,
-            condExpr
-        );
-
-        // if (!truthy) break;
-        var breakIfFalse = ExprTree.IfThen(
-            ExprTree.Not(truthyCall),
-            ExprTree.Break(breakLabel)
-        );
-
-        // Loop body
-        var loopBody = ExprTree.Block(
-            typeof(void),
-            breakIfFalse,
-            bodyExpr
-        );
-
-        // Expression.Loop with break and continue labels
-        return ExprTree.Block(
-            typeof(object),
-            ExprTree.Loop(loopBody, breakLabel, continueLabel),
-            ExprTree.Constant(null, typeof(object))
-        );
-    }
-
-    /// <summary>
     /// foreach文の変換
     /// foreach (item in collection) { body }
     /// </summary>
@@ -568,6 +520,138 @@ public class CodeGenerator
             // loop
             ExprTree.Loop(loopBody, breakLabel, continueLabel),
             // return null
+            ExprTree.Constant(null, typeof(object))
+        );
+    }
+
+    /// <summary>
+    /// for文の変換
+    /// パターン1: for (item in collection) { body } - コレクション反復
+    /// パターン2: for (condition) { body } - 条件ループ
+    /// </summary>
+    private ExprTree GenerateForStmt(ForStmt stmt)
+    {
+        return stmt.Kind switch
+        {
+            ForStmtKind.Collection => GenerateForCollection(stmt),
+            ForStmtKind.Condition => GenerateForCondition(stmt),
+            _ => throw new NotImplementedException($"Unknown ForStmtKind: {stmt.Kind}")
+        };
+    }
+
+    /// <summary>
+    /// for (item in collection) { body } の変換
+    /// foreach と同じ実装
+    /// </summary>
+    private ExprTree GenerateForCollection(ForStmt stmt)
+    {
+        var breakLabel = ExprTree.Label($"forBreak_{_labelCounter}");
+        var continueLabel = ExprTree.Label($"forContinue_{_labelCounter++}");
+
+        // コレクション式を評価
+        var collectionExpr = GenerateExpression(stmt.Collection!);
+
+        // IEnumerable にキャスト
+        var enumerableVar = ExprTree.Variable(typeof(System.Collections.IEnumerable), "enumerable");
+        var enumeratorVar = ExprTree.Variable(typeof(System.Collections.IEnumerator), "enumerator");
+        var currentVar = ExprTree.Variable(typeof(object), "current");
+
+        // ループ変数をctx.Globalsに登録
+        var setLoopVar = ExprTree.Call(
+            ExprTree.Property(_ctxParam, nameof(ScriptContext.Globals)),
+            typeof(Dictionary<string, object?>).GetMethod("set_Item")!,
+            ExprTree.Constant(stmt.IteratorVariable),
+            currentVar
+        );
+
+        // ループラベルをスタックにプッシュ
+        _loopLabels.Push((breakLabel, continueLabel));
+
+        var bodyExpr = GenerateExpression(stmt.Body);
+
+        // ループラベルをポップ
+        _loopLabels.Pop();
+
+        // ループ本体
+        var loopBody = ExprTree.Block(
+            typeof(void),
+            // if (!enumerator.MoveNext()) break;
+            ExprTree.IfThen(
+                ExprTree.Not(ExprTree.Call(enumeratorVar, typeof(System.Collections.IEnumerator).GetMethod("MoveNext")!)),
+                ExprTree.Break(breakLabel)
+            ),
+            // current = enumerator.Current
+            ExprTree.Assign(currentVar, ExprTree.Property(enumeratorVar, nameof(System.Collections.IEnumerator.Current))),
+            // ctx.Globals[variable] = current
+            setLoopVar,
+            // body
+            bodyExpr
+        );
+
+        // for全体
+        return ExprTree.Block(
+            typeof(object),
+            new[] { enumerableVar, enumeratorVar, currentVar },
+            // enumerable = (IEnumerable)collection
+            ExprTree.Assign(
+                enumerableVar,
+                ExprTree.TypeAs(collectionExpr, typeof(System.Collections.IEnumerable))
+            ),
+            // enumerator = enumerable.GetEnumerator()
+            ExprTree.Assign(
+                enumeratorVar,
+                ExprTree.Call(enumerableVar, typeof(System.Collections.IEnumerable).GetMethod("GetEnumerator")!)
+            ),
+            // loop
+            ExprTree.Loop(loopBody, breakLabel, continueLabel),
+            // return null
+            ExprTree.Constant(null, typeof(object))
+        );
+    }
+
+    /// <summary>
+    /// for (condition) { body } の変換
+    /// while と同じ実装
+    /// </summary>
+    private ExprTree GenerateForCondition(ForStmt stmt)
+    {
+        var breakLabel = ExprTree.Label($"forBreak_{_labelCounter}");
+        var continueLabel = ExprTree.Label($"forContinue_{_labelCounter++}");
+
+        // ループラベルをスタックにプッシュ
+        _loopLabels.Push((breakLabel, continueLabel));
+
+        var condExpr = GenerateExpression(stmt.Condition!);
+        var bodyExpr = GenerateExpression(stmt.Body);
+
+        // ループラベルをポップ
+        _loopLabels.Pop();
+
+        // IsTruthy で真偽値判定
+        var truthyCall = ExprTree.Call(
+            typeof(RuntimeHelpers),
+            nameof(RuntimeHelpers.IsTruthy),
+            null,
+            condExpr
+        );
+
+        // if (!truthy) break;
+        var breakIfFalse = ExprTree.IfThen(
+            ExprTree.Not(truthyCall),
+            ExprTree.Break(breakLabel)
+        );
+
+        // Loop body
+        var loopBody = ExprTree.Block(
+            typeof(void),
+            breakIfFalse,
+            bodyExpr
+        );
+
+        // Expression.Loop with break and continue labels
+        return ExprTree.Block(
+            typeof(object),
+            ExprTree.Loop(loopBody, breakLabel, continueLabel),
             ExprTree.Constant(null, typeof(object))
         );
     }
@@ -1048,6 +1132,27 @@ public class CodeGenerator
             nameof(RuntimeHelpers.CreateHash),
             Type.EmptyTypes,
             pairArrayExpr
+        );
+    }
+
+    /// <summary>
+    /// 範囲リテラルの変換
+    /// 仕様: 1..10 → RuntimeHelpers.CreateRange(1, 10, false)
+    ///       1...10 → RuntimeHelpers.CreateRange(1, 10, true)
+    /// </summary>
+    private ExprTree GenerateRangeExpr(RangeExpr expr)
+    {
+        var startExpr = GenerateExpression(expr.Start);
+        var endExpr = GenerateExpression(expr.End);
+        var inclusiveExpr = ExprTree.Constant(expr.Inclusive);
+
+        return ExprTree.Call(
+            typeof(RuntimeHelpers),
+            nameof(RuntimeHelpers.CreateRange),
+            Type.EmptyTypes,
+            startExpr,
+            endExpr,
+            inclusiveExpr
         );
     }